sample_apps/minimal/requirements.txt
python-dotenv
fastapi
tortoise-api-model
tortoise-orm[asyncpg]
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart